refactor standalone pipeline to separate files in src and root
