+++ .pre-commit-config.yaml
@@ -9,6 +9,11 @@ repos:
       - id: end-of-file-fixer
       # Check YAML files are valid
       - id: check-yaml
+        exclude: |
+          (?x)^(
+            standalone_utils/.*\.yaml|
+            standalone_utils/.*\.yml
+          )$
       # Check TOML files are valid
       - id: check-toml
       # Check JSON files are valid
@@ -38,7 +43,7 @@ repos:
         additional_dependencies:
           - tomli
         args: [
-          '--ignore-words-list=nd,te,ue,hist,tha,manuel,fpr',  # Common false positives in scientific code
+          '--ignore-words-list=nd,te,ue,hist,tha,manuel,fpr,mape',  # Common false positives in scientific code
           '--skip=*.lock,*.json',  # Skip lock files and data files
         ]
