+++ setup.sh
@@ -1,2 +1,3 @@
+#!/bin/bash
 source .venv/bin/activate
 # module load tensorflow
