+++ src/hep_foundation/data/physlite_percentiles.json
@@ -5,9 +5,9 @@
     "timestamp": "2025-08-04T17:30:55.867811"
   },
   "Track d0": {
-    "0.1": -15.783986587524414,
-    "99.9": 10.89181268692019,
-    "timestamp": "2025-08-07T09:48:40.260903"
+    "0.1": -17.487511863708495,
+    "99.9": 11.046281476976302,
+    "timestamp": "2025-08-07T11:06:45.833237"
   },
   "Track z0": {
     "0.1": -142.1888956604004,
