+++ src/hep_foundation/standalone_models/standalone_dnn_regressor.py
