+++ src/hep_foundation/standalone_models/standalone_plot_manager.py
